# lib/api-zod/src/generated/types/incidentUpdateStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NullTrace AI DevOps Observability Platform API
 * OpenAPI spec version: 0.1.0
 */

export type IncidentUpdateStatus = typeof IncidentUpdateStatus[keyof typeof IncidentUpdateStatus];


export const IncidentUpdateStatus = {
  OPEN: 'OPEN',
  INVESTIGATING: 'INVESTIGATING',
  RESOLVED: 'RESOLVED',
  CLOSED: 'CLOSED',
} as const;