# lib/api-zod/src/generated/types/podHealth.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NullTrace AI DevOps Observability Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PodHealthStatus } from './podHealthStatus';

export interface PodHealth {
  id: number;
  name: string;
  namespace: string;
  status: PodHealthStatus;
  service: string;
  cpu: number;
  memory: number;
  restarts: number;
  age: string;
  node: string;
}